feat: 체력 회복 기능 추가

+++ Fantasy-Grower/Assets/Scripts/Core/EntityStatModifier.cs
@@ -1,10 +1,10 @@
 ﻿[System.Serializable]
 public struct EntityStatModifier
 {
-    public int BonusHp;
+    public float BonusHp;
     public float BonusHpRecovery;
     public float BonusDamageReduction;
-    public int BonusAttackPower;
+    public float BonusAttackPower;
     public float BonusAttackSpeed;
     public float BonusCriticalPercentage;
